requirements.txt
# Core API & Web
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# LangGraph & Agents
langgraph>=0.3.0
langchain>=0.3.0
langchain-core>=0.3.50

# Multi-LLM Providers
langchain-google-genai>=4.2.0   # Google Gemini (primary)
langchain-openai>=0.3.0         # OpenRouter (uses OpenAI-compatible API)
langchain-mistralai>=0.2.0      # Mistral AI
langchain-huggingface>=0.1.0    # Hugging Face Inference

# Data & Validation
pydantic>=2.10.0
pydantic-settings>=2.5.0

# File & PDF
PyPDF2>=3.0.0

# Environment & Utils
python-dotenv>=1.0.0
typing-extensions>=4.12.0
tenacity>=9.0.0
loguru>=0.7.0
httpx>=0.27.0

# Redis + Cache
redis>=5.0.0

# Background tasks
celery>=5.4.0

# Job Scraping
python-jobspy>=1.1.0
beautifulsoup4>=4.12.3
requests>=2.32.0

# Data processing
pandas>=2.1.0
numpy>=1.26.0

# Database persistence
langgraph-checkpoint-postgres>=3.0.0
psycopg[binary]>=3.3.0

